`hstore` column mapping support
It is now possible to specify a `hstore` column mapping as the fourth
argument to `copy_fields_from_foreign_table()` and
`fallback_to_fields_from_foreign_table()`, so that the local and foreign
fields do not need to have identical names.

I also fixed the feature that allows you to altogether omit the column
names, and made sure that this is now covered by tests.